requirements.txt
streamlit>=1.32
beautifulsoup4
lxml
requests
pandas
Pillow
